+++ src/Akka.sln
@@ -2,8 +2,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.22823.1
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30723.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Examples", "Examples", "{69279534-1DBA-4115-BF8B-03F77FC8125E}"
 EndProject
@@ -205,6 +203,7 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Cluster.Sharding.Tests", "contrib\cluster\Akka.Cluster.Sharding.Tests\Akka.Cluster.Sharding.Tests.csproj", "{AD74F85A-B641-4C70-9427-1B822209A8DC}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.Cluster.Tools.Tests", "contrib\cluster\Akka.Cluster.Tools.Tests\Akka.Cluster.Tools.Tests.csproj", "{1E0D56C4-19DA-4BEF-BED8-EB13F1A48CB1}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.TestKit.Xunit2", "contrib\testkits\Akka.TestKit.Xunit2\Akka.TestKit.Xunit2.csproj", "{7DBD5C17-5E9D-40C4-9201-D092751532A7}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.MultiNodeTests", "core\Akka.MultiNodeTests\Akka.MultiNodeTests.csproj", "{F0781BEA-5BA0-4AF0-BB15-E3F209B681F5}"
@@ -215,9 +214,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.DI.StructureMap", "con
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BasicStructureMapUses", "contrib\dependencyInjection\Examples\BasicStructureMapUses\BasicStructureMapUses.csproj", "{13BA1CC1-A431-441D-8B11-3969D2C68A6E}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ClusterSharding", "ClusterSharding", "{D8F27F25-4EC3-4746-B3A8-BC94413A6830}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.ClusterSharding", "contrib\clustersharding\Akka.ClusterSharding\Akka.ClusterSharding.csproj", "{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Cluster", "Cluster", "{32EF87DD-22FA-4825-89E5-CB88E4082C6B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -800,30 +797,6 @@ Global
 		{13BA1CC1-A431-441D-8B11-3969D2C68A6E}.Release Mono|Any CPU.Build.0 = Release|Any CPU
 		{13BA1CC1-A431-441D-8B11-3969D2C68A6E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{13BA1CC1-A431-441D-8B11-3969D2C68A6E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63}.Debug Mono|Any CPU.ActiveCfg = Debug|Any CPU
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63}.Debug Mono|Any CPU.Build.0 = Debug|Any CPU
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63}.Release Mono|Any CPU.ActiveCfg = Release|Any CPU
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63}.Release Mono|Any CPU.Build.0 = Release|Any CPU
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Debug Mono|Any CPU.ActiveCfg = Debug|Any CPU
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Debug Mono|Any CPU.Build.0 = Debug|Any CPU
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Release Mono|Any CPU.ActiveCfg = Release|Any CPU
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Release Mono|Any CPU.Build.0 = Release|Any CPU
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug Mono|Any CPU.ActiveCfg = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug Mono|Any CPU.Build.0 = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release Mono|Any CPU.ActiveCfg = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release Mono|Any CPU.Build.0 = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -904,7 +877,10 @@ Global
 		{4B693CFF-F14F-4924-95A1-5FA896D749D9} = {D1CCD86E-0EF8-473A-979B-25E1235FEA2D}
 		{D1CCD86E-0EF8-473A-979B-25E1235FEA2D} = {B1D10183-8FAE-4506-B935-403FCED89BDB}
 		{32EABA12-DDAA-4F2A-B254-85239267D869} = {D1CCD86E-0EF8-473A-979B-25E1235FEA2D}
-		{13BA1CC1-A431-441D-8B11-3969D2C68A6E} = {D1CCD86E-0EF8-473A-979B-25E1235FEA2D}
+		{2065C3A2-8C15-4912-BCF5-AE89E3DDA079} = {B1D10183-8FAE-4506-B935-403FCED89BDB}
+		{BEF84A6F-32C4-4ACF-AFC3-7B5FCA6F209E} = {7625FD95-4B2C-4A5B-BDD5-94B1493FAC8E}
+		{D63223FA-03F5-4B32-A6EC-668F718C0826} = {7625FD95-4B2C-4A5B-BDD5-94B1493FAC8E}
+		{264C22A4-CAFC-41F6-B82C-4DDC5C196767} = {588C1513-FAB6-42C3-B6FC-3485F13620CF}
 		{BAC85686-AFC4-413E-98DC-5ED8F468BC63} = {264C22A4-CAFC-41F6-B82C-4DDC5C196767}
 		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE} = {264C22A4-CAFC-41F6-B82C-4DDC5C196767}
 		{52999BCC-7C31-4D38-92E9-B564077294A9} = {588C1513-FAB6-42C3-B6FC-3485F13620CF}
@@ -912,18 +888,10 @@ Global
 		{7AE6DE46-24A7-41EB-A18C-5C3AB748E81E} = {52999BCC-7C31-4D38-92E9-B564077294A9}
 		{AD74F85A-B641-4C70-9427-1B822209A8DC} = {52999BCC-7C31-4D38-92E9-B564077294A9}
 		{1E0D56C4-19DA-4BEF-BED8-EB13F1A48CB1} = {52999BCC-7C31-4D38-92E9-B564077294A9}
-		{2065C3A2-8C15-4912-BCF5-AE89E3DDA079} = {B1D10183-8FAE-4506-B935-403FCED89BDB}
-		{BEF84A6F-32C4-4ACF-AFC3-7B5FCA6F209E} = {7625FD95-4B2C-4A5B-BDD5-94B1493FAC8E}
-		{D63223FA-03F5-4B32-A6EC-668F718C0826} = {7625FD95-4B2C-4A5B-BDD5-94B1493FAC8E}
-		{264C22A4-CAFC-41F6-B82C-4DDC5C196767} = {588C1513-FAB6-42C3-B6FC-3485F13620CF}
 		{7DBD5C17-5E9D-40C4-9201-D092751532A7} = {7625FD95-4B2C-4A5B-BDD5-94B1493FAC8E}
 		{F0781BEA-5BA0-4AF0-BB15-E3F209B681F5} = {01167D3C-49C4-4CDE-9787-C176D139ACDD}
 		{3B9E6211-9488-4DB5-B714-24248693B38F} = {264C22A4-CAFC-41F6-B82C-4DDC5C196767}
 		{34E5B4E5-0ED0-4A27-B53A-BFD812D45E1E} = {B1D10183-8FAE-4506-B935-403FCED89BDB}
 		{13BA1CC1-A431-441D-8B11-3969D2C68A6E} = {D1CCD86E-0EF8-473A-979B-25E1235FEA2D}
-		{BAC85686-AFC4-413E-98DC-5ED8F468BC63} = {264C22A4-CAFC-41F6-B82C-4DDC5C196767}
-		{5A3C24D7-0D1C-4974-BBB4-22AC792666DE} = {264C22A4-CAFC-41F6-B82C-4DDC5C196767}
-		{D8F27F25-4EC3-4746-B3A8-BC94413A6830} = {588C1513-FAB6-42C3-B6FC-3485F13620CF}
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4} = {D8F27F25-4EC3-4746-B3A8-BC94413A6830}
 	EndGlobalSection
 EndGlobal
