--- src/contrib/clustersharding/Akka.ClusterSharding.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Akka.ClusterSharding", "Akka.ClusterSharding\Akka.ClusterSharding.csproj", "{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2FF1A911-5E34-46DB-B2A8-6AFB65259FB4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
